Week 14 updates

+++ src/main/java/pet/store/entity/Employee.java
@@ -17,7 +17,7 @@ public class Employee {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer employeeId;
+	private Long employeeId;
 	
 	private String employeeFirstName;
 	private String employeeLastName;
