+++ src/main/java/pet/store/entity/PetStore.java
@@ -23,7 +23,7 @@ public class PetStore {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer petStoreId;
+	private Long petStoreId;
 
 	private String petStoreName;
 	private String petStoreAddress;
